+++ ERPManagementAPP/Maintain/OperatingMaintainControl.cs
@@ -58,6 +58,9 @@ namespace ERPManagementAPP.Maintain
                 Refresh_Main_GridView();
             }
             action.Commands.Add(FlyoutCommand.Yes);
+            Delectaction.Commands.Add(FlyoutCommand.Yes);
+            Delectaction.Commands.Add(FlyoutCommand.Cancel);
+            Delectaction.Caption = "刪除確認";
             #region 營運資料表
             gridView1.OptionsSelection.EnableAppearanceFocusedCell = false;
             #region 營運資訊報表按鈕
@@ -223,19 +226,23 @@ namespace ERPManagementAPP.Maintain
             btn_Operating_Delete.Click += (s, e) =>
             {
                 FocuseMainGrid();
-                string response = APIMethod.Delete_Operating(FocuseOperatingMainSetting.OperatingFlag, FocuseOperatingMainSetting.OperatingNumber);
-                if (response == "200")
+                Delectaction.Description = $"刪除編碼 : {FocuseOperatingMainSetting.OperatingNumber}";
+                if (FlyoutDialog.Show(Form1, Delectaction) == DialogResult.Yes)
                 {
-                    Refresh_Main_GridView();
-                    action.Caption = "刪除營運資訊成功";
-                    action.Description = "";
-                    FlyoutDialog.Show(Form1, action);
-                }
-                else
-                {
-                    action.Caption = "刪除營運資訊失敗";
-                    action.Description = "";
-                    FlyoutDialog.Show(Form1, action);
+                    string response = APIMethod.Delete_Operating(FocuseOperatingMainSetting.OperatingFlag, FocuseOperatingMainSetting.OperatingNumber);
+                    if (response == "200")
+                    {
+                        Refresh_Main_GridView();
+                        action.Caption = "刪除營運資訊成功";
+                        action.Description = "";
+                        FlyoutDialog.Show(Form1, action);
+                    }
+                    else
+                    {
+                        action.Caption = "刪除營運資訊失敗";
+                        action.Description = "";
+                        FlyoutDialog.Show(Form1, action);
+                    }
                 }
             };
             #endregion
