+++ ERPManagementAPP/Maintain/PickingMaintainControl.cs
@@ -61,6 +61,9 @@ namespace ERPManagementAPP.Maintain
                 Refresh_Main_GridView();
             }
             action.Commands.Add(FlyoutCommand.Yes);
+            Delectaction.Commands.Add(FlyoutCommand.Yes);
+            Delectaction.Commands.Add(FlyoutCommand.Cancel);
+            Delectaction.Caption = "刪除確認";
             #region 領料資料表
             gridView1.OptionsSelection.EnableAppearanceFocusedCell = false;
             #region 領料資訊報表按鈕
@@ -226,19 +229,23 @@ namespace ERPManagementAPP.Maintain
             btn_Picking_Delete.Click += (s, e) =>
             {
                 FocuseMainGrid();
-                string response = APIMethod.Delete_Picking(FocusePickingMainSetting.PickingFlag, FocusePickingMainSetting.PickingNumber);
-                if (response == "200")
+                Delectaction.Description = $"刪除編碼 : {FocusePickingMainSetting.PickingNumber}";
+                if (FlyoutDialog.Show(Form1, Delectaction) == DialogResult.Yes)
                 {
-                    Refresh_Main_GridView();
-                    action.Caption = "刪除領料資訊成功";
-                    action.Description = "";
-                    FlyoutDialog.Show(Form1, action);
-                }
-                else
-                {
-                    action.Caption = "刪除領料資訊失敗";
-                    action.Description = "";
-                    FlyoutDialog.Show(Form1, action);
+                    string response = APIMethod.Delete_Picking(FocusePickingMainSetting.PickingFlag, FocusePickingMainSetting.PickingNumber);
+                    if (response == "200")
+                    {
+                        Refresh_Main_GridView();
+                        action.Caption = "刪除領料資訊成功";
+                        action.Description = "";
+                        FlyoutDialog.Show(Form1, action);
+                    }
+                    else
+                    {
+                        action.Caption = "刪除領料資訊失敗";
+                        action.Description = "";
+                        FlyoutDialog.Show(Form1, action);
+                    }
                 }
             };
             #endregion
