+++ ERPManagementAPP/Maintain/PurchaseMaintainControl.cs
@@ -56,6 +56,9 @@ namespace ERPManagementAPP.Maintain
                 Refresh_Main_GridView();
             }
             action.Commands.Add(FlyoutCommand.Yes);
+            Delectaction.Commands.Add(FlyoutCommand.Yes);
+            Delectaction.Commands.Add(FlyoutCommand.Cancel);
+            Delectaction.Caption = "刪除確認";
             #region 進貨資料表
             gridView1.OptionsSelection.EnableAppearanceFocusedCell = false;
             #region 進貨資訊報表按鈕
@@ -221,19 +224,23 @@ namespace ERPManagementAPP.Maintain
             btn_Purchase_Delete.Click += (s, e) =>
             {
                 FocuseMainGrid();
-                string response = APIMethod.Delete_Purchase(FocusePurchaseMainSetting.PurchaseFlag, FocusePurchaseMainSetting.PurchaseNumber);
-                if (response == "200")
+                Delectaction.Description = $"刪除編碼 : {FocusePurchaseMainSetting.PurchaseNumber}";
+                if (FlyoutDialog.Show(Form1, Delectaction) == DialogResult.Yes)
                 {
-                    Refresh_Main_GridView();
-                    action.Caption = "刪除進貨資訊成功";
-                    action.Description = "";
-                    FlyoutDialog.Show(Form1, action);
-                }
-                else
-                {
-                    action.Caption = "刪除進貨資訊失敗";
-                    action.Description = "";
-                    FlyoutDialog.Show(Form1, action);
+                    string response = APIMethod.Delete_Purchase(FocusePurchaseMainSetting.PurchaseFlag, FocusePurchaseMainSetting.PurchaseNumber);
+                    if (response == "200")
+                    {
+                        Refresh_Main_GridView();
+                        action.Caption = "刪除進貨資訊成功";
+                        action.Description = "";
+                        FlyoutDialog.Show(Form1, action);
+                    }
+                    else
+                    {
+                        action.Caption = "刪除進貨資訊失敗";
+                        action.Description = "";
+                        FlyoutDialog.Show(Form1, action);
+                    }
                 }
             };
             #endregion
